requirements.txt
asgiref @ file:///C:/b/abs_ae1n0p_8_w/croot/asgiref_1724072476516/work
asttokens==2.4.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
executing==2.0.1
gunicorn==23.0.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==26.1.1
setuptools==75.1.0
six==1.16.0
sqlparse @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/sqlparse_1699544474746/work
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
tzdata @ file:///croot/python-tzdata_1690578112552/work
wcwidth==0.2.13
wheel==0.44.0
whitenoise==6.8.2
